Split @controller
